requirements.txt
airportsdata
fastapi
httpx
pandas
psycopg2-binary
python-dotenv
sqlalchemy
streamlit
uvicorn
